add Logic and View calsses
